requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
iniconfig==1.1.1
Markdown==3.4.1
packaging==22.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.0
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
psycopg2-binary==2.9.3
